refactor: Create Utils folder and separate CircularList from Node classes
